requirements.txt
ib_async>=1.0.0
py-clob-client>=0.34
anthropic>=0.40
python-telegram-bot==20.7
aiohttp>=3.9
pyyaml>=6.0
python-dotenv>=1.0
numpy>=1.26
pandas>=2.1
scipy>=1.12
statsmodels>=0.14
aiosqlite>=0.19
feedparser>=6.0
